Rename video element to standard element
