+++ src/index.ts
@@ -8,7 +8,7 @@ export type {
 export { pullquoteElement } from "./elements/pullquote/PullquoteSpec";
 export { codeElement } from "./elements/code/CodeElementSpec";
 export { createImageElement } from "./elements/image/ImageElement";
-export { createVideoElement } from "./elements/video/VideoSpec";
+export { createStandardElement } from "./elements/standard/StandardSpec";
 export { richlinkElement } from "./elements/rich-link/RichlinkSpec";
 export { createInteractiveElement } from "./elements/interactive/InteractiveSpec";
 export {
